Updated cuda - 13.3.0-1

+++ world/cuda/PKGBUILD
@@ -4,11 +4,11 @@
 
 pkgname=(cuda)
 pkgbase=cuda
-pkgver=13.2.1
+pkgver=13.3.0
 # Before upgrading, make sure that we ship at least this version of
 # nvidia-utils as otherwise we'll get stuff such as #7.
-_driverver=595.58.03
-pkgrel=3
+_driverver=610.43.02
+pkgrel=1
 pkgdesc="NVIDIA's GPU programming toolkit"
 arch=(x86_64 aarch64)
 url="https://developer.nvidia.com/cuda-zone"
@@ -83,8 +83,8 @@ sha512sums=('ff6ee01a6838e51804911c50f817c8e1545d71aa1b34cd937a546d42642097f81d5
             '5df94583c7d082045fffb5c055b163b5242e7975d4de19e334c2d4b2e0e8caf8a930a5ff82b800f54d45ac53d9be0044afd893fb649b461594981693a5256836'
             '200b23fa74c486a3b1d003d3326163a7d2318dd3a02a0fe488fef3c4ae713cd0dac251b751edc723696fb3d04d4e5f2c11d3b68a0b5f4621388ca28a387ba421'
             '0a49ce7972ac24ac1f1bd229a4fad3a37747d3e8123c4c1c4521c19fbc308b85f3ee6bf8fbee420aac2335c57aa6b524c20277c59abd7202590144f7ed79c24a')
-sha512sums_x86_64=('2746aac0f1843995672409666536471b94d69d127f488693b79b4b0758606267978930533ff832d5e26b698bf1a85e9e19e01424f003630b96866b2cf142b673')
-sha512sums_aarch64=('467316f6b1a539d519a2a1f05fc7d55e829becb0b8615b63c23c93eb602379926fc1e01d55cd4e09619a5043268a96c9579e15fa84bbf3524f37d45edc85bedf')
+sha512sums_x86_64=('9e882eba72b4a986f9ab25afb9566d54f8d73c88aa0634a6e3cd986f05c4c90fd6c4a8cc55f87fb4cedc0bda3a560079f9856a075c5c61b3b773787cc89792fe')
+sha512sums_aarch64=('14dfc609f987177ea4bab2b66468d17cf75c8f340392151352c03843ed17d62d98e416e5b642df21ae6d31a6ef48002f1fda515024d5f70905f44307896fd02b')
 
 prepare() {
   sh cuda_${pkgver}_${_driverver}_linux*.run --target "${srcdir}" --noexec
@@ -125,12 +125,15 @@ build() {
     rm -r "$lib"
   done
 
-  # Replace CCCL headers with symlinks to the cccl package
-  rm -r "${_prepdir}"/opt/cuda/targets/x86_64-linux/include/cccl/
-  ln -s /usr/include/cccl "${_prepdir}"/opt/cuda/targets/x86_64-linux/include/cccl
-  rm -r "${_prepdir}"/opt/cuda/targets/x86_64-linux/lib/cmake/{cccl,cub,libcudacxx,thrust}
-  rmdir "${_prepdir}"/opt/cuda/targets/x86_64-linux/lib/cmake/
-  ln -s /usr/lib/cmake "${_prepdir}"/opt/cuda/targets/x86_64-linux/lib/cmake
+  # no cccl for aarch64
+  if [[ $CARCH == "x86_64"* ]]; then
+    # Replace CCCL headers with symlinks to the cccl package
+    rm -r "${_prepdir}"/opt/cuda/targets/x86_64-linux/include/cccl/
+    ln -s /usr/include/cccl "${_prepdir}"/opt/cuda/targets/x86_64-linux/include/cccl
+    rm -r "${_prepdir}"/opt/cuda/targets/x86_64-linux/lib/cmake/{cccl,cub,libcudacxx,thrust}
+    rmdir "${_prepdir}"/opt/cuda/targets/x86_64-linux/lib/cmake/
+    ln -s /usr/lib/cmake "${_prepdir}"/opt/cuda/targets/x86_64-linux/lib/cmake
+  fi
 
   # Fix location of CUPTI headers and libs, otherwise some software fails to build
   # https://gitlab.archlinux.org/archlinux/packaging/packages/tensorflow/-/work_items/19#note_273055
